+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+-   [xy] basic signal-processing props (convolution, downsampling) for curves,
+    areas, and intervals
 -   [band] allow `BandSurface` to appear using tooltip context
 -   [core] component `AxisTooltip`
 -   [examples] bar chart with gradients within bars